+++ spec/Definition/ObjectDefinitionSpec.php
@@ -60,6 +60,12 @@ class ObjectDefinitionSpec extends ObjectBehavior
         $this->getDefinitionData()->shouldHaveArgumentsEquals(['hello', 'world']);
     }
 
+    function it_can_define_constructor_arguments_with_old_method()
+    {
+        $this->setConstructArgs(['hello', 'world']);
+        $this->getDefinitionData()->shouldHaveArgumentsEquals(['hello', 'world']);
+    }
+
     function it_can_define_a_method_call_with_arguments()
     {
         $this->call('doSomething')->with('hello', 'world');
@@ -70,6 +76,18 @@ class ObjectDefinitionSpec extends ObjectBehavior
                 'arguments' => ['hello', 'world']
             ]
 
+        );
+    }
+
+    function it_can_define_a_method_call_with_arguments_using_old_method()
+    {
+        $this->setMethod('doSomething', ['hello', 'world']);
+        $this->getDefinitionData()->shouldHaveACallEquals(
+            [
+                'type' => DefinitionData::METHOD,
+                'name' => 'doSomething',
+                'arguments' => ['hello', 'world']
+            ]
         );
     }
 
@@ -92,11 +110,38 @@ class ObjectDefinitionSpec extends ObjectBehavior
         );
     }
 
+    function it_can_define_an_assignment_to_a_property_with_old_method()
+    {
+        $this->setProperty('scope', 'test');
+        $this->getDefinitionData()->shouldHaveACallEquals(
+            [
+                'type' => DefinitionData::PROPERTY,
+                'name' => 'scope',
+                'arguments' => 'test'
+            ]
+        );
+    }
+
     function it_throws_property_not_found_exception_if_defining_an_undefined_property_call()
     {
         $this->beConstructedWith(InitializableService::class);
         $this->shouldThrow(PropertyNotFoundException::class)
             ->during('to', ['test']);
+    }
+
+    function it_can_chain_multiple_method_calls()
+    {
+        $this->call('doSomething')->with('hello1');
+        $this->call('doSomething')->with('hello2');
+        $this->call('doSomething')->with('hello3');
+        $this->getDefinitionData()->shouldHaveACallEquals(
+            [
+                'type' => DefinitionData::METHOD,
+                'name' => 'doSomething',
+                'arguments' => ['hello1']
+            ]
+
+        );
     }
 
     public function getMatchers()